# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lembur Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/post-employees": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Post Employees",
                "description": "Post Employees",
                "operationId": "post-employees",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "salary",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful stored"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get Employees information",
                "description": "Returns Employees data",
                "operationId": "getEmployees",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/App/Http/Resources/EmployeesResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/post-overtimes": {
            "post": {
                "tags": [
                    "postOvertimes"
                ],
                "summary": "Post Overtimes",
                "description": "Post Overtimes",
                "operationId": "post-overtimes",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "time_started",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "time_ended",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful stored"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    }
}